Renamed import to convert string to module
The function allTestData, that receives strings that names the module, will import the modules and pass it along using the module identifier.
The problem is that some identifier inside the  module might override and module name. In the snippet `import yada;` it is not guaranteed that `yada` identifier represents the module, it is possible that a identifier from within the module is named `yada` also.
By using renamed imports, it is possible to give an unique and non-ambiguous name for each imported module.
Remarks: Using static import old yield issues with packages, thats why renaming is prefered.