Struct Mock doesn't import modules
Changes how the struct Mock declare the return type and type of the parameters for an overriden of a function so that the module that contains such types may be not imported.
Aliases in the generated mixin are created to identify the types using traits on the identifier T which is the class that is being mocked.