Allow more than 26 temporary directories on Windows
The current implementation using mktemp sets five of the Xs
to the thread ID, and only uses the other letter in sequence.
This limits it to 26 directories per template. This change
throws in additional randomness in the template in an effort
to expand that limit, more like we have on Linux, while still
using mktemp for the actual final name.