Unify the way file / line arguments are passed
In some parts of the code, file was passed as 'in', and in others as 'const'.
The former is obviously wrong, as a 'in' should imply 'scope',
and message is stored in an 'Exception', which by definition escapes the scope.
Additionally, 'const' in itself isn't very useful: since the elements
are immutable anyway, a 'const string' implicityle converts to a 'string'.
Last but not least, 'line' was passed as 'in' in some places.
For consistency, and because 'in' had no effect, it was removed.